% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_data.R
\name{balance}
\alias{balance}
\title{Use expected level means to balance an unbalanced design}
\usage{
balance(data, level_means, global_mean = zero_vector(data))
}
\arguments{
\item{data}{An object of class \code{nesteddata}. This must only be unbalanced at the individual level,
with dummy levels present to represent unobserved levels.}

\item{level_means}{A list of matrices. The names of the list elements must be the same as those of
the group factors in data. Each list element should be a matrix whose row names are the levels of
the corresponding factor and whose rows are the relevant mean values.}

\item{global_mean}{A numeric vector. The global mean of the model.}
}
\value{
An object of class \code{nesteddata}, which is a balanced version of the input \code{data}.
}
\description{
Use expected level means to balance an unbalanced design
}
