% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_checks.R
\name{converged_l_infinity}
\alias{converged_l_infinity}
\title{Generate a closure that checks for convergence of matrices in the l-infinity norm}
\usage{
converged_l_infinity(eps = 1e-06)
}
\arguments{
\item{eps}{The threshold at which convergence is reaced}
}
\description{
Generate a closure that checks for convergence of matrices in the l-infinity norm
}
