% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{degf}
\alias{degf}
\title{Degrees of freedom for nested designs}
\usage{
degf(ns, factor_idx, fixed_intercept = FALSE)
}
\arguments{
\item{ns}{Ordered vector of number of levels per factor}

\item{factor_idx}{Index of factor for which to compute degrees of freedom}

\item{fixed_intercept}{Logical indicating whether the model includes a fixed intercept}
}
\description{
Computes I\link{k} ... I\link{n-1} (I\link{n} - 1), or I\link{n} if k = n in the fixed-intercept case.
}
