% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesteddata.R
\name{dim.nesteddata}
\alias{dim.nesteddata}
\title{Dimension of the response variable of a nested data object}
\usage{
\method{dim}{nesteddata}(x)
}
\description{
Dimension of the response variable of a nested data object
}
