% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_algebra.R
\name{eigensqrt}
\alias{eigensqrt}
\title{Use eigen to compute a square root of the non-negative definite symmetric matrix A}
\usage{
eigensqrt(A, compute_inverse = FALSE)
}
\arguments{
\item{A}{Non-negative definite symmetric matrix}

\item{compute_inverse}{Logical indicating if inverse should also be computed. If this is \code{TRUE}, \code{A}
must be positive definite.}
}
\value{
If \code{compute_inverse} is \code{FALSE}, a matrix with dimensions matching \code{A}. If it is \code{TRUE}, a
list with entries names \code{sqrt} and \code{invsqrt}, containing a square root and inverse square root of
\code{A} respectively.
}
\description{
The resulting U may not be symmetric, but satisfies \verb{U'U == A}
}
