% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nesteddata.R
\name{fit_balanced_nesteddata}
\alias{fit_balanced_nesteddata}
\title{Computes REML or ML estimates of nested covariance components}
\usage{
fit_balanced_nesteddata(sos, constraints, data, ...)
}
\arguments{
\item{sos}{List of sum-of-squares matrices}

\item{constraints}{Matrix of order constraints for the factors}

\item{data}{An object inheriting \code{nesteddata}}

\item{...}{Other arguments to pass to \code{multiordered_wishart_ml}}
}
\value{
A list of estimated covariance components
}
\description{
Computes REML or ML estimates of nested covariance components
}
