% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nesteddata.R
\name{init_covs_nesteddata}
\alias{init_covs_nesteddata}
\title{Produces a list of identity matrices to use as initial guesses for covariance components}
\usage{
init_covs_nesteddata(data)
}
\arguments{
\item{data}{An object inheriting \code{nesteddata}}
}
\value{
A named list of identity matrices with an entry corresponding to each factor in \code{data}
}
\description{
Produces a list of identity matrices to use as initial guesses for covariance components
}
