% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{linear_order}
\alias{linear_order}
\title{Generate pairwise constraints representing a linear order}
\usage{
linear_order(v)
}
\arguments{
\item{v}{A character vector of variable names in increasing order.}
}
\description{
Generate pairwise constraints representing a linear order
}
