% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_wishart_ml.R
\name{multiordered_wishart_ml}
\alias{multiordered_wishart_ml}
\title{ML estimates for Wishart covariance under general order}
\usage{
multiordered_wishart_ml(
  A_list,
  constraints,
  max.iter = 1000L,
  check_convergence = converged_l_infinity(1e-06)
)
}
\arguments{
\item{A_list}{A names list of square matrices with a \code{degf} attribute.}

\item{constraints}{A 2-row character matrix specifying the order constraints. Each column \code{v} encodes
a single order constraint through \code{v[1] <= v[2]}.}

\item{max.iter}{The number of EM iterations before termination.}
}
\description{
Given a named list of realisations of independent Wishart matrices, compute the maximum likelihood
estimates of their underlying covariances under a given set of binary order constraints.
}
\details{
This is done using an algorithm due to Calvin and Dykstra (1991)
}
