% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesteddata.R
\name{nesteddata}
\alias{nesteddata}
\title{Create a \code{nesteddata} object from a data frame}
\usage{
nesteddata(df, factors)
}
\arguments{
\item{df}{Wide data frame with one column for each level and one for each dimension of the
response variable.}

\item{factors}{Character vector containing the names of the level variables. Must be supplied in
the bottom-up order of nesting so that \code{factor[i]} is nested inside \code{factor[i+1]}.}
}
\value{
An object of class \code{nesteddata} containing the information necessary to compute covariance
component estimates.
The underlying object is a list with entries
\itemize{
\item \verb{$sos_matrix} The matrix X'X, where X is the matrix of responses.
\item \verb{$group_sums} Matrix of sums of responses grouped by the second-lowest factor.
}

The object also contains the following attributes:
\itemize{
\item \code{attr(, "n_factors")} Integer recording the number of factors in the model. This will be 2 and
3 for 1- and 2-way designs respectively.
\item \code{attr(, "factors")} Character vector of the names of the factors. The order determines the
nesting structure, with the factor \code{attr(, "factors")[i]} nested inside \code{attr(, "factors")[i+1]}.
\item \code{attr(, "parents")} List of character vectors indexed by the names of all factors except the
lowest and highest. The string \code{attr(, "parents")$factor["level"]} is the name of the level that
"level" is nested inside.
\item \code{attr(, "n_levels")} Integer vector recording the maximum number of different levels nested
in each parent for each factor.
\item \code{attr(, "n_observed")} List of integer vectors for every factor except the lowest. Records
the number of levels observed one level below each factor level.
}
}
\description{
Create a \code{nesteddata} object from a data frame
}
