% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_wishart_ml.R
\name{ordered_wishart_ml}
\alias{ordered_wishart_ml}
\title{ML estimates for Wishart covariance under linear order}
\usage{
ordered_wishart_ml(A1, A2)
}
\arguments{
\item{A1, A2}{Square matrices with a \code{degf} attribute.}
}
\value{
A list with entries \code{primal} and \code{dual}
}
\description{
Given realisations of independent Wishart matrices A1 and A2, compute maximum likelihood estimates of
their underlying covariances Sigma1 and Sigma2 under the restriction that 0 <= Sigma1 <= Sigma2 with
respect to the non-negative definite ordering.
}
\details{
This is done using an algorithm due to Amemiya (1985).
}
