% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_algebra.R
\name{paired_inverse}
\alias{paired_inverse}
\title{Compute (A^(-1) + n E^(-1))^(-1) for a vector of ns}
\usage{
paired_inverse(A, E, E_type = c("cov", "prec"))
}
\arguments{
\item{A}{Symmetric nonnegative-definite square matrix}

\item{E}{Symmetric positive-definite square matrix}

\item{E_type}{If this equals \code{"prec"}, the \code{E} argument is instead \code{E^(-1)}}

\item{ns}{A vector of numerics}
}
\value{
A list of inverted matrices indexed by the vector \code{ns}
}
\description{
Compute (A^(-1) + n E^(-1))^(-1) for a vector of ns
}
