% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnesteddata.R
\name{rnesteddata}
\alias{rnesteddata}
\title{Sample a random balanced nested design}
\usage{
rnesteddata(
  covs,
  nlevels_per_factor,
  global_mean = rep(0, nrow(covs[[1]])),
  factors = names(covs),
  level_names = NULL,
  var_names = rownames(covs[[1]]),
  format = c("table", "data")
)
}
\arguments{
\item{covs}{List of covariance components}

\item{nlevels_per_factor}{Named numeric vector indicating the number of distinct levels at each
level of nestine}

\item{factors}{The names of the factors in increasing order of the nesting structure}

\item{level_names}{A list of level names for each factor}

\item{var_names}{A character vector of the names of the variables}

\item{format}{A character vector indicating the type of outpu}

\item{gloal_mean}{Numeric vector. The fixed effect of the model}
}
\value{
Either a tibble or an object of class \code{nesteddata}, depending on the value of \code{format}.
}
\description{
Sample a random balanced nested design
}
