% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{sumofsquares}
\alias{sumofsquares}
\title{Compute sum-of-squares matrices from nested data}
\usage{
sumofsquares(data, fixed_intercept = FALSE)
}
\arguments{
\item{data}{An object of class \code{nesteddata}}

\item{fixed_intercept}{Logical indicating whether the global mean in the model is a fixed effect.
This should be \code{TRUE} for maximum likelihood estimation. If \code{TRUE}, \code{A[n]} will have an additional
degree of freedom.}
}
\value{
A list of matrices indexed by factor name. Each matrix has a \code{degf} attribute, which encodes
the degrees of freedom in the corresponding likelihood function. This is usually the degrees of
freedom of the Wishart distribution, except in the case of ML estimation.
}
\description{
Computes a list of between-factor sum-of-squares matrices (A\link{1}, A\link{2}, ..., A\link{n}).
If the model is correct, these are independent and have a Wishart distribution with
}
\details{
A\link{k} ~ W(degf = (I\link{k} - 1) I\link{k} ... I\link{n},
cov  = Sigma\link{1} + I\link{1} Sigma\link{2} + ... + I\link{1} ... I\link{n-1} Sigma\link{n}))
}
