% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{weight}
\alias{weight}
\title{Number of individuals per level at a given factor}
\usage{
weight(ns, factor_idx)
}
\arguments{
\item{ns}{Ordered vector of number of levels per factor}

\item{factor_idx}{Index of factor for which to compute weight}
}
\description{
Computes (I\link{1} ... I\link{k-1})
}
