% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_data.R
\name{extend_names}
\alias{extend_names}
\title{Create a vector of new names that don't conflict with existing names}
\usage{
extend_names(n_missing, existing_names = character(0))
}
\arguments{
\item{n_missing}{Named integer vector detailing how many new names are to be created per parent.}

\item{existing_names}{Character vector of names that already exist, and so must be avoided.}
}
\description{
Create a vector of new names that don't conflict with existing names
}
