% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_algebra.R
\name{paired_inverse}
\alias{paired_inverse}
\title{Compute (A^(-1) + n E^(-1))^(-1) for a vector of ns}
\usage{
paired_inverse(A, E, ns, keep_names = TRUE)
}
\arguments{
\item{A}{Symmetric nonnegative-definite covariance matrix}

\item{E}{Symmetric positive-definite covariance or precision matrix}

\item{ns}{A vector of numerics}
}
\value{
A list of inverted matrices indexed by the vector \code{ns}
}
\description{
Compute (A^(-1) + n E^(-1))^(-1) for a vector of ns
}
