% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{unbalanced_EM}
\alias{unbalanced_EM}
\title{Uses the EM algorithm to maximise a given criterion in unbalanced data}
\usage{
unbalanced_EM(
  crit_argmax,
  sos_conditioner,
  init_params,
  max.iter = 100L,
  check_convergence = function(curr_params, prev_params) FALSE,
  crit_argmax_params = list(),
  sos_conditioner_params = list()
)
}
\arguments{
\item{crit_argmax}{An function that maximises a criterion of interest given a list of arguments}

\item{init_params}{A list of initial parameter guesses}

\item{max.iter}{maximum iterations for the EM algorithm}

\item{check_convergence}{a function taking as arguments the current and previous parameters that
retugns a logical indicating whether the EM algorithm has converged.}

\item{crit_argmax_params}{List of non-updating arguments to pass to \code{crit_argmax}}

\item{sos_conditioner_params}{List of non-updating arguments to pass to \code{sos_conditioner}}

\item{arg_conditioner}{A function that computes the conditional expectation of sum-of-squares
matrices required for the \code{crit_argmax} function. Should take as its first argument the updating
#' parameter values.}
}
\value{
A list of parameters in the format of \code{init_params}
}
\description{
Uses the EM algorithm to maximise a given criterion in unbalanced data
}
